+++ collective/googleanalytics/viewlets/tracking.py
@@ -1,3 +1,5 @@
+import sys
+from urllib import urlencode
 from zope.component import queryMultiAdapter
 from Products.CMFCore.utils import getToolByName
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
@@ -6,7 +8,7 @@ from collective.googleanalytics.interfaces.tracking import IAnalyticsTrackingPlu
 
 class AnalyticsTrackingViewlet(AnalyticsViewlet):
     """
-    A viewlet that inserts the Google Analytics tracking code 
+    A viewlet that inserts the Google Analytics tracking code
     at the end of the page. We override the default Plone viewlet
     so that we can exclude the code for certain roles.
     """
@@ -23,14 +25,14 @@ class AnalyticsTrackingViewlet(AnalyticsViewlet):
         Checks to see whether the viewlet should be rendered based on the role
         of the user and the selections for excluded roles in the configlet.
         """
-                
+
         member = self.membership_tool.getAuthenticatedMember()
-        
+
         for role in self.analytics_tool.tracking_excluded_roles:
             if member.has_role(role):
                 return False
         return True
-        
+
     def getTrackingWebProperty(self):
         """
         Returns the Google web property ID for the selected tracking profile,
@@ -38,13 +40,13 @@ class AnalyticsTrackingViewlet(AnalyticsViewlet):
         """
 
         return self.analytics_tool.__dict__.get('tracking_web_property', None)
-        
+
     def renderPlugins(self):
         """
         Render each of the selected tracking plugins for the current context
         and request.
         """
-        
+
         results = []
         for plugin_name in self.analytics_tool.tracking_plugin_names:
             plugin = queryMultiAdapter(
@@ -57,3 +59,23 @@ class AnalyticsTrackingViewlet(AnalyticsViewlet):
                 results.append(plugin())
         return '\n'.join(results)
 
+    def getsearchcat(self):
+        return self.view.__name__
+
+    def renderPageview(self):
+        push_params = ["'_trackPageview'"]
+        exc_info = sys.exc_info()[0]
+        if self.view.__name__.startswith("search"):
+            query = {'q': self.request.get('SearchableText', ''),
+                     'searchcat': self.getsearchcat()}
+            push_params.append("'/searchresult?%s'" % urlencode(query))
+        elif exc_info is not None:
+            if exc_info == NotFound:
+                errorcode = 404
+            else:
+                errorcode = 500
+            push_params.append(
+            ("'/error/%s?page=' + document.location.pathname + "
+             "document.location.search + '&from=' + document.referrer")
+            % errorcode)
+        return "_gaq.push([%s]);" % ', '.join(push_params)